feat/final touhes

+++ buttons.js
@@ -4,17 +4,36 @@ if (JSON.parse(localStorage.getItem("tasks")).length > 0) {
 
         if (evento.target && evento.target.id.startsWith("deleteTask_")) {
 
-            let indiceDaTarefa = evento.target.id.split("_")[1];
+            Swal.fire({
+                title: "Queres mesmo apagar esta tarefa?",
+                text: "Não poderás voltar atrás!",
+                icon: "warning",
+                showCancelButton: true,
+                confirmButtonColor: "#3085d6",
+                cancelButtonColor: "#d33",
+                confirmButtonText: "APAGA!"
+                }).then((result) => {
+                if (result.isConfirmed){
+                    let indiceDaTarefa = evento.target.id.split("_")[1];
 
-            let listaDeTarefas = JSON.parse(localStorage.getItem("tasks")) || [];
+                    let listaDeTarefas = JSON.parse(localStorage.getItem("tasks")) || [];
 
-            listaDeTarefas.splice(indiceDaTarefa, 1);
+                    listaDeTarefas.splice(indiceDaTarefa, 1);
 
-            localStorage.setItem("tasks", JSON.stringify(listaDeTarefas));
+                    localStorage.setItem("tasks", JSON.stringify(listaDeTarefas));
 
-            console.log("Tarefa apagada:", indiceDaTarefa);
+                    Swal.fire({
+                        title: "Apagado!",
+                        text: "Tarefa apagada com sucesso.",
+                        icon: "success"
+                    }).then(() => {
+                        window.location.reload();
+                    });
+                } 
 
-            window.location.reload(true);
+            });
+
+
         }
 
         if (evento.target && evento.target.id.startsWith("finishTask_")) {
@@ -35,7 +54,14 @@ if (JSON.parse(localStorage.getItem("tasks")).length > 0) {
 
             console.log("Tarefa Completa:", indiceDaTarefa);
 
-            window.location.reload(true);
+            Swal.fire({
+            title: "Bom trabalho :)!",
+            text: "Terminaste a tarefa!",
+            icon: "success"
+            }).then(() => {
+                window.location.reload();
+            });
+
         }
 
         if (evento.target && evento.target.id.startsWith("editTask_")) {
@@ -51,7 +77,6 @@ if (JSON.parse(localStorage.getItem("tasks")).length > 0) {
                 <th><button id="saveTask_${indice}">save</button></th>
                 <th><input id="editTitle_${indice}" value="${tarefa.title}"></th>
                 <th><input id="editText_${indice}" value="${tarefa.text}"></th>
-                <th><input id="editDate_${indice}" value="${tarefa.dateOfCreation}"></th>
                 <th>
                     <button id="cancelEdit_${indice}">cancel</button>
                 </th>
@@ -63,14 +88,14 @@ if (JSON.parse(localStorage.getItem("tasks")).length > 0) {
             let indice = evento.target.id.split("_")[1];
             let tasks = JSON.parse(localStorage.getItem("tasks")) || [];
 
+            let data = JSON.parse(localStorage.getItem("tasks.dateOfCreation")) || [];
             let novoTitulo = document.getElementById(`editTitle_${indice}`).value;
             let novoTexto = document.getElementById(`editText_${indice}`).value;
-            let novaData = document.getElementById(`editDate_${indice}`).value;
 
             tasks[indice] = {
                 title: novoTitulo,
                 text: novoTexto,
-                dateOfCreation: novaData
+                dateOfCreation: tasks[indice].dateOfCreation
             };
 
             localStorage.setItem("tasks", JSON.stringify(tasks));
